requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
idna==3.2
pycparser==2.20
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
